net, tests, udn: Centralize UDN fixtures and single_nic marker
- Move fixtures from test module to conftest.
- Centralize single_nic marker on the class.
- Use IPERF_SERVER_PORT.
