+++ tests/network/user_defined_network/conftest.py
@@ -5,7 +5,9 @@ from kubernetes.dynamic import DynamicClient
 from ocp_resources.user_defined_network import Layer2UserDefinedNetwork
 
 from libs.net.ip import random_ipv4_address
+from libs.net.traffic_generator import IPERF_SERVER_PORT, TcpServer, VMTcpClient
 from libs.net.udn import UDN_BINDING_DEFAULT_PLUGIN_NAME
+from libs.net.vmspec import lookup_iface_status_ip, lookup_primary_network
 from libs.vm import affinity
 from libs.vm.oper import run_vms
 from libs.vm.vm import BaseVirtualMachine
@@ -43,6 +45,54 @@ def namespaced_layer2_user_defined_network(admin_client, udn_namespace):
 @pytest.fixture(scope="module")
 def udn_affinity_label():
     return affinity.new_label(key_prefix="udn")
+
+
+@pytest.fixture(scope="class")
+def vma_udn(udn_namespace, namespaced_layer2_user_defined_network, udn_affinity_label, admin_client):
+    with udn_vm(
+        namespace_name=udn_namespace.name,
+        name="vma-udn",
+        client=admin_client,
+        binding=UDN_BINDING_DEFAULT_PLUGIN_NAME,
+        template_labels=dict((udn_affinity_label,)),
+    ) as vm:
+        vm.start(wait=True)
+        vm.wait_for_agent_connected()
+        yield vm
+
+
+@pytest.fixture(scope="class")
+def vmb_udn(udn_namespace, namespaced_layer2_user_defined_network, udn_affinity_label, admin_client):
+    with udn_vm(
+        namespace_name=udn_namespace.name,
+        name="vmb-udn",
+        client=admin_client,
+        binding=UDN_BINDING_DEFAULT_PLUGIN_NAME,
+        template_labels=dict((udn_affinity_label,)),
+    ) as vm:
+        vm.start(wait=True)
+        vm.wait_for_agent_connected()
+        yield vm
+
+
+@pytest.fixture(scope="class")
+def server(vmb_udn):
+    with TcpServer(vm=vmb_udn, port=IPERF_SERVER_PORT) as server:
+        assert server.is_running()
+        yield server
+
+
+@pytest.fixture(scope="class")
+def client(vma_udn, vmb_udn):
+    with VMTcpClient(
+        vm=vma_udn,
+        server_ip=str(
+            lookup_iface_status_ip(vm=vmb_udn, iface_name=lookup_primary_network(vm=vmb_udn).name, ip_family=4)
+        ),
+        server_port=IPERF_SERVER_PORT,
+    ) as client:
+        assert client.is_running()
+        yield client
 
 
 @pytest.fixture(scope="class")
